requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
black==24.1.1
click==8.1.7
databases==0.8.0
fastapi==0.109.2
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
isort==5.13.2
logging==0.4.9.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.7
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
